Fundamental Kotlin - Return at labels.
