--- CodeExamples/Fundamental-Kotlin/src/main/java/net/milosvasic/fundamental/kotlin/basics/returns/ReturnAtLabels.kt
@@ -1,34 +1,0 @@
-package net.milosvasic.fundamental.kotlin.basics.returns
-
-fun returns1(ints: List<Int>, toBreakAt: Int) {
-    ints.forEach {
-        item ->
-        if (item == toBreakAt) return
-        println("Item: $item")
-    }
-}
-
-fun returns2(ints: List<Int>, toBreakAt: Int) {
-    ints.forEach myLabel1@ {
-        item ->
-        if (item % 2 == 0) return@myLabel1
-        println("Item: $item")
-    }
-}
-
-fun returns3(ints: List<Int>, toBreakAt: Int) {
-    ints.forEach {
-        item ->
-        if (item % 2 == 0) return@forEach
-        println("Item: $item")
-    }
-}
-
-fun returns4(ints: List<Int>, toBreakAt: Int) {
-    ints.forEach(
-            fun(value: Int) {
-                if (value % 2 == 0) return
-                println("Item: $value")
-            }
-    )
-}
